+++ tests/formatter/input-output-pairs/multiline_match_statements.in.gd
@@ -35,9 +35,9 @@ class X:
 				pass
 			var xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx:
 				pass
-			# TODO: format to multiple lines once multiline syntax is fixed in Godot
 			[_,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1]:
 				pass
-			# TODO: format to multiple lines once multiline syntax is fixed in Godot
 			{'aaaaa', 'bbbbbbbbbbbbbbbbbbbbb': 1111, 'ccccccccccccccccccccc': 1111, 'ddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd': 1111}:
+				pass
+			[1,2,]:
 				pass